Add files via upload
Embark on a journey through the bustling streets of Namma Bengaluru as we delve into the chaos and order of its infamous traffic. Bangalore, the Garden City of India, is no stranger to the relentless buzz of vehicles, traffic jams, and unfortunate accidents. However, amidst this seemingly chaotic scenario, is there an underlying pattern waiting to be discovered? In this exploration, we turn our attention to the 'Bangalore-road-data' dataset sourced from Kaggle. This dataset serves as our guide to unravel the mysteries of Bangalore's traffic.